refactor: factor out generic MediaController
